+++ src/twitch/handler/SoundSettingsPopup.hpp
@@ -21,6 +21,7 @@ protected:
     void onSoundSelect(CCObject *);
     void onPlaySound(CCObject *sender);
     void onClose(CCObject *);
+    void onOpenCustomSfx(CCObject *);
 
 public:
     static SoundSettingsPopup *create(CommandSettingsPopup *parent, int actionIdx, const std::string &selectedSound, std::function<void(const std::string &)> onSave);