+++ changelog.md
@@ -8,6 +8,7 @@
 - Replaced the input fonts to **Pusab**
 - Added Time value on the **Notification** action settings.
 - Added Copy Action
+- Added **Custom SFX** support for the **Sound Effect** action.
 
 # v0.1.10-beta.1
 - Added **Experimental Features** option in the Mod Settings.